int->double for time-related variables
git-svn-id: file:///srv/local/data/cleanRepo/trunk@27456 a0517364-e873-4167-b1c8-416daa5a6b91

+++ contribs/dvrp/src/main/java/org/matsim/contrib/dvrp/VrpSimEngine.java
@@ -94,7 +94,7 @@ public class VrpSimEngine
     }
 
 
-    public void nextTask(Schedule<? extends Task> schedule, int time)
+    public void nextTask(Schedule<? extends Task> schedule, double time)
     {
         optimizer.nextTask(schedule);
         notifyAgentLogics();
