+++ contribs/dvrp/src/main/java/org/matsim/contrib/dvrp/data/schedule/Schedules.java
@@ -59,7 +59,7 @@ public class Schedules
     };
 
 
-    public static int getActualT1(Schedule<?> schedule)
+    public static double getActualT1(Schedule<?> schedule)
     {
         Vehicle veh = schedule.getVehicle();
